+++ fj-doc-lib-autodoc/src/main/java/org/fugerit/java/doc/lib/autodoc/facade/XsdParserFacade.java
@@ -1,12 +1,8 @@
 package org.fugerit.java.doc.lib.autodoc.facade;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import org.fugerit.java.core.cfg.ConfigException;
 import org.fugerit.java.doc.lib.autodoc.parser.model.AutodocModel;
 import org.xmlet.xsdparser.core.XsdParser;
-import org.xmlet.xsdparser.xsdelements.XsdElement;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -15,17 +11,20 @@ public class XsdParserFacade {
 
 	public AutodocModel parse(String filePath) throws ConfigException {
 		return ConfigException.getWithMessage( () -> {
-			AutodocModel autodocModel = null;
 			XsdParser xsdParser = new XsdParser(filePath);
-			autodocModel = new AutodocModel( xsdParser );
-			List<XsdElement> elements = xsdParser.getResultXsdElements().collect(Collectors.toList());
-			for (XsdElement currentElement : elements) {
-				log.info("current element {} -> {}", currentElement.getName(), currentElement.getType());
+			final AutodocModel autodocModel = new AutodocModel( xsdParser );
+			xsdParser.getResultXsdElements().forEach( currentElement -> {
+				log.debug("current element {} -> {}", currentElement.getName(), currentElement.getType());
 				autodocModel.addElement( currentElement );
-			}
+			} );
+			xsdParser.getResultXsdSchemas().forEach( currentSchema -> {
+				log.debug("current schema {} -> {}", currentSchema.getTargetNamespace(), currentSchema.getNamespaces() );
+				currentSchema.getNamespaces().entrySet().forEach( e -> log.debug( "namespace : {} : {}", e.getKey(), e.getValue().getName() ) );
+				currentSchema.getChildrenComplexTypes().forEach( autodocModel::addType );
+				currentSchema.getChildrenSimpleTypes().forEach( autodocModel::addSimpleType );
+			} );
 			return autodocModel;
 		} , "Error parsing xsd" );
 	}
-
 
 }