+++ fj-doc-lib-autodoc/src/main/java/org/fugerit/java/doc/lib/autodoc/parser/model/AutodocModel.java
@@ -5,9 +5,15 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 
+import lombok.Getter;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
 import org.xmlet.xsdparser.core.XsdParser;
+import org.xmlet.xsdparser.xsdelements.XsdComplexType;
 import org.xmlet.xsdparser.xsdelements.XsdElement;
+import org.xmlet.xsdparser.xsdelements.XsdSimpleType;
 
+@Slf4j
 public class AutodocModel implements Serializable {
 
 	public static final String ATT_NAME = "autodocModel";
@@ -17,63 +23,54 @@ public class AutodocModel implements Serializable {
 	 */
 	private static final long serialVersionUID = 5016692762844545990L;
 
-	private LinkedHashMap<String, AutodocElement> elements;
+	private final LinkedHashMap<String, AutodocElement> elements;
 
-	private transient XsdParser xsdParser;
+	private final LinkedHashMap<String, AutodocType> types;
+
+	private final LinkedHashMap<String, AutodocSimpleType> simpleTypes;
+
+	@Getter
+	private final transient XsdParser xsdParser;
+
+	@Getter @Setter
+	private String version;
+
+	@Getter @Setter
+	private String title;
+
+	@Getter @Setter
+	private String xsdPrefix;
+
+	@Getter @Setter
+	private String autodocPrefix;
 	
 	public AutodocModel( XsdParser xsdParser ) {
 		this.elements = new LinkedHashMap<>();
+		this.types = new LinkedHashMap<>();
+		this.simpleTypes = new LinkedHashMap<>();
 		this.xsdParser = xsdParser;
-	}
-	
-	private String version;
-	
-	private String title;
-	
-	private String xsdPrefix;
-	
-	private String autodocPrefix;
-	
-	public String getVersion() {
-		return version;
-	}
-
-	public void setVersion(String version) {
-		this.version = version;
-	}
-
-	public String getTitle() {
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-	}
-
-	public String getXsdPrefix() {
-		return xsdPrefix;
-	}
-
-	public void setXsdPrefix(String xsdPrefix) {
-		this.xsdPrefix = xsdPrefix;
-	}
-
-	public String getAutodocPrefix() {
-		return autodocPrefix;
-	}
-
-	public void setAutodocPrefix(String autodocPrefix) {
-		this.autodocPrefix = autodocPrefix;
-	}
-
-	public XsdParser getXsdParser() {
-		return xsdParser;
 	}
 
 	public AutodocElement addElement( XsdElement xsdElement ) {
 		AutodocElement element = new AutodocElement( this, xsdElement );
 		this.elements.put( element.getKey(), element );
 		return element;
+	}
+
+	public AutodocType addType( XsdComplexType xsdComplexType ) {
+		AutodocType type = new AutodocType(  xsdComplexType );
+		String key = AutodocUtils.toKey( xsdComplexType );
+		log.debug( "xsdComplexType key : {}", key );
+		this.types.put( key, type );
+		return type;
+	}
+
+	public AutodocSimpleType addSimpleType( XsdSimpleType xsdSimpleType ) {
+		AutodocSimpleType type = new AutodocSimpleType(  this, xsdSimpleType );
+		String key = AutodocUtils.toKey( xsdSimpleType );
+		log.debug( "xsdSimpleType key : {}", key );
+		this.simpleTypes.put( key, type );
+		return type;
 	}
 	
 	public Collection<String> getElementNames() {
@@ -91,5 +88,13 @@ public class AutodocModel implements Serializable {
 	public boolean containsName( String name ) {
 		return this.elements.containsKey( name );
 	}
-	
+
+	public Collection<AutodocType> getTypes() {
+		return Collections.unmodifiableCollection( this.types.values() );
+	}
+
+	public Collection<AutodocSimpleType> getSimpleTypes() {
+		return Collections.unmodifiableCollection( this.simpleTypes.values() );
+	}
+
 }